+++ src/app/mocks/browser.ts
@@ -6,11 +6,13 @@ import { likeHandlers } from './handler/like';
 import { followHandler } from './handler/follow';
 import { searchHandler } from './handler/search';
 import { userHandler } from './handler/user';
+import { bookmarkHandlers } from './handler/bookmark';
 
 // 브라우저에서 실행하기 위한 mocking worker 초기화
 export const worker = setupWorker(
   ...commentHandlers,
   ...feedHandlers,
+  ...bookmarkHandlers,
   ...likeHandlers,
   ...followHandler,
   ...searchHandler,
